requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.7.0; python_version >= '3.9'
certifi==2024.12.14; python_version >= '3.6'
click==8.1.8; python_version >= '3.7'
distro==1.9.0; python_version >= '3.6'
fastapi==0.115.6; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jiter==0.8.2; python_version >= '3.8'
openai==1.58.1; python_version >= '3.8'
pydantic==2.10.4; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
python-telegram-bot==21.9; python_version >= '3.9'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.41.3; python_version >= '3.8'
telegram==0.0.1
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
uvicorn==0.34.0; python_version >= '3.9'
